futures_trader.py
import ccxt.pro as ccxtpro
import asyncio
import logging
import os
from dotenv import load_dotenv
from strategies.bos_fvg_strategy import BOSFVGStrategy

# --- Setup ---
load_dotenv()
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s',
    handlers=[logging.StreamHandler()]
)
logger = logging.getLogger(__name__)

# --- Configuration ---
EXCHANGE_NAME = "binance"  # Placeholder - e.g., 'binance', 'bybit'
SYMBOL = "BTC/USDT:USDT"  # Placeholder for a futures contract
TIMEFRAME = "15m"

# --- Main Trading Logic ---
async def main():
    """The main function to run the futures trading bot."""
    logger.info(f"Starting futures trader for {SYMBOL} on {EXCHANGE_NAME}")

    # 1. Connect to the exchange
    api_key = os.getenv(f"{EXCHANGE_NAME.upper()}_API_KEY")
    api_secret = os.getenv(f"{EXCHANGE_NAME.upper()}_SECRET")

    if not api_key or not api_secret:
        logger.error(f"API keys for {EXCHANGE_NAME.upper()} not found in .env file. Exiting.")
        return

    exchange_class = getattr(ccxtpro, EXCHANGE_NAME)
    exchange = exchange_class({
        'apiKey': api_key,
        'secret': api_secret,
        'options': {
            'defaultType': 'future',
        },
    })

    try:
        # Set sandbox mode if available
        if exchange.has.get('setSandboxMode'):
            exchange.set_sandbox_mode(True)
            logger.info("Sandbox mode enabled.")

        logger.info(f"Successfully connected to {exchange.id}")

        # 2. Fetch real-time data using websockets
        while True:
            try:
                logger.info(f"Watching for OHLCV data for {SYMBOL}...")
                # Placeholder for websocket loop
                # In a real implementation, we would process candles here
                await asyncio.sleep(10) # Keep alive

            except Exception as e:
                logger.error(f"An error occurred in the main loop: {e}")
                break

    finally:
        await exchange.close()
        logger.info("Exchange connection closed.")

if __name__ == "__main__":
    try:
        asyncio.run(main())
    except KeyboardInterrupt:
        logger.info("Trader stopped manually.")
