requirements.txt
# Core Dependencies
pandas>=1.3.0
numpy>=1.21.0
pandas-ta>=0.3.14b0
matplotlib>=3.4.0
seaborn>=0.11.0
yfinance>=0.1.70
tqdm>=4.62.0
ccxt>=2.0.0
python-dotenv>=0.19.0
